Prepare for next refactor
